refactor(linter): move logic for combining `JsFix`es into separate function (#19052)
Pure refactor. Move logic for combining multiple `JsFix`es into a single `Fix` into a separate function. This function can be reused in `RuleTester` for testing fixes.
